requirements.txt
fastapi
uvicorn
requests
python-dotenv
ollama
pydantic
